Changed menu behavior on scroll / added more content
Before, there was a waypoint buggy JQuery sort of control over the page
scroll. When the page reached a certain element it applied some classes
to a target element - the header, in this case. Now, i let the header
be position:absolute to scroll with the page and a new directive
listens to the page scroll in a better fashion in order to determine
when to apply the position:fixed to the header-menu.

Also added some content and adjusted its behaviour to medium and small
sized IOS devices. I’ve used flex to deliver responsivity, but i’m
still not happy with its ease of use and browser support.