Added angular-translate support
Angular-translate was hooked to the system and it’s being correctly
updated.

I’ve also created a messages file app/pages/messages.js with some
content which is binded on our Menu and on our Home;

However, there’s TONS of texts to be dumped there in both languages.

Lots of sweat to come. It works!